+++ hal/src/main/native/systemcore/CAN.cpp
@@ -6,24 +6,237 @@
 
 #include "hal/Errors.h"
 
+#include "wpinet/EventLoopRunner.h"
+#include "wpi/mutex.h"
+#include "wpi/DenseMap.h"
+
+#include "wpinet/uv/Poll.h"
+#include "wpinet/uv/Timer.h"
+
+#include <net/if.h>
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <sys/ioctl.h>
+
 #include <linux/can.h>
 #include <linux/can/raw.h>
 
+#include "hal/Threads.h"
+
+#define NUM_CAN_BUSES 1
+
+namespace {
+
+struct FrameStore {
+  canfd_frame frame;
+  uint64_t timestamp{0};
+};
+
+struct SocketCanState {
+  wpi::EventLoopRunner loopRunner;
+  wpi::mutex writeMutex[NUM_CAN_BUSES];
+  int socketHandle[NUM_CAN_BUSES];
+  wpi::mutex timerMutex;
+  // ms to count/timer map
+  wpi::DenseMap<uint16_t, std::pair<size_t, std::weak_ptr<wpi::uv::Timer>>>
+      timers;
+  // ms to bus mask/packet
+  wpi::DenseMap<uint16_t, std::vector<std::pair<uint8_t, canfd_frame>>>
+      timedFrames;
+  // packet to time
+  wpi::DenseMap<uint32_t, uint16_t> packetToTime;
+
+  wpi::mutex readMutex[NUM_CAN_BUSES];
+  // TODO, we need a MUCH better way of doing this masking
+  wpi::DenseMap<uint32_t, FrameStore> readFrames[NUM_CAN_BUSES];
+
+  bool InitializeBuses();
+};
+
+}  // namespace
+
+static SocketCanState* canState;
+
 namespace hal::init {
-void InitializeCAN() {}
+void InitializeCAN() {
+  canState = new SocketCanState{};
+}
 }  // namespace hal::init
+
+bool SocketCanState::InitializeBuses() {
+  bool success = true;
+  loopRunner.ExecSync([this, &success](wpi::uv::Loop& loop) {
+    int32_t status = 0;
+    HAL_SetCurrentThreadPriority(true, 50, &status);
+    if (status != 0) {
+      printf("Failed to set CAN thread priority\n");
+    }
+
+    for (int i = 0; i < NUM_CAN_BUSES; i++) {
+      std::scoped_lock lock{writeMutex[i]};
+      socketHandle[i] = socket(PF_CAN, SOCK_RAW, CAN_RAW);
+      if (socketHandle[i] == -1) {
+        success = false;
+        return;
+      }
+
+      ifreq ifr;
+      strcpy(ifr.ifr_name, "canX");
+      ifr.ifr_name[3] = ('0' + i);
+
+      if (ioctl(socketHandle[i], SIOCGIFINDEX, &ifr) == -1) {
+        success = false;
+        return;
+      }
+
+      sockaddr_can addr;
+      memset(&addr, 0, sizeof(addr));
+      addr.can_family = AF_CAN;
+      addr.can_ifindex = ifr.ifr_ifindex;
+
+      if (bind(socketHandle[i], reinterpret_cast<const sockaddr*>(&addr),
+               sizeof(addr)) == -1) {
+        success = false;
+        return;
+      }
+
+      printf("Successfully bound to can interface %d\n", i);
+
+      auto poll = wpi::uv::Poll::Create(loop, socketHandle[i]);
+      if (!poll) {
+        success = false;
+        return;
+      }
+
+      poll->pollEvent.connect([fd = socketHandle[i]](int mask) {
+        if (mask & UV_READABLE) {
+          canfd_frame frame;
+          int rVal = read(fd, &frame, sizeof(frame));
+          if (rVal <= 0) {
+            // TODO error handling
+            return;
+          }
+          if (frame.can_id & CAN_ERR_FLAG) {
+            // Do nothing if this is an error frame
+            return;
+          }
+        }
+      });
+
+      poll->Start(UV_READABLE);
+    }
+  });
+  return success;
+}
+
+namespace hal {
+bool InitializeCanBuses() {
+  return canState->InitializeBuses();
+}
+}  // namespace hal
+
+namespace {
+
+static constexpr uint32_t MatchingBitMask = CAN_EFF_MASK | CAN_RTR_FLAG;
+
+static_assert(CAN_RTR_FLAG == HAL_CAN_IS_FRAME_REMOTE);
+static_assert(CAN_EFF_FLAG == HAL_CAN_IS_FRAME_11BIT);
+
+uint32_t MapMessageIdToSocketCan(uint32_t id) {
+  // Message and RTR map directly
+  uint32_t toRet = id & MatchingBitMask;
+
+  // Reverse the 11 bit flag
+  if ((id & HAL_CAN_IS_FRAME_11BIT) == 0) {
+    toRet |= CAN_EFF_FLAG;
+  }
+}
+}  // namespace
 
 extern "C" {
 
 void HAL_CAN_SendMessage(uint32_t messageID, const uint8_t* data,
                          uint8_t dataSize, int32_t periodMs, int32_t* status) {
-  *status = HAL_HANDLE_ERROR;
-  return;
+  // TODO this will become a parameter
+  // isFd will also be a part of this parameter
+  uint8_t busId = 0;
+
+  if (busId >= NUM_CAN_BUSES) {
+    *status = PARAMETER_OUT_OF_RANGE;
+    return;
+  }
+
+  bool isFd = false;
+  messageID = MapMessageIdToSocketCan(messageID);
+
+  canfd_frame frame;
+  memset(&frame, 0, sizeof(frame));
+  frame.can_id = messageID;
+  frame.flags = isFd ? CANFD_FDF | CANFD_BRS : 0;
+  if (dataSize) {
+    auto size = (std::min)(dataSize, static_cast<uint8_t>(sizeof(frame.data)));
+    memcpy(frame.data, data, size);
+    frame.len = size;
+  }
+
+  if (periodMs == HAL_CAN_SEND_PERIOD_NO_REPEAT) {
+    auto mtu = isFd ? CANFD_MTU : CAN_MTU;
+    std::scoped_lock lock{canState->writeMutex[busId]};
+    int result = send(canState->socketHandle[busId], &frame, mtu, 0);
+    if (result == mtu) {
+      *status = 0;
+    } else {
+      // TODO better error
+      *status = HAL_ERR_CANSessionMux_InvalidBuffer;
+    }
+  } else {
+    // TODO add periodic support
+    *status = HAL_ERR_CANSessionMux_NotAllowed;
+    return;
+  }
 }
 void HAL_CAN_ReceiveMessage(uint32_t* messageID, uint32_t messageIDMask,
                             uint8_t* data, uint8_t* dataSize,
                             uint32_t* timeStamp, int32_t* status) {
-  *status = HAL_HANDLE_ERROR;
+  uint8_t busId = 0;
+
+  if (busId >= NUM_CAN_BUSES) {
+    *status = PARAMETER_OUT_OF_RANGE;
+    return;
+  }
+
+  std::scoped_lock lock{canState->readMutex[busId]};
+
+  // TODO, this is going to be wrong, but we're going to assume that
+  // any lookup without the 11 bit mask set wants to look for a 29 bit frame.
+  // Also, only do fast lookups for 29 bit frames
+
+  // Fast case is the following.
+  // Mask doesn't include 11 bit flag
+  // Mask doesn't include RTR flag
+  // Mask is full
+  if (messageIDMask == CAN_EFF_MASK) {
+    // We're doing a fast lookup
+    auto& msg = canState->readFrames[busId][*messageID];
+    if (msg.timestamp == 0) {
+      *status = HAL_ERR_CANSessionMux_MessageNotFound;
+      return;
+    }
+    if (*dataSize < msg.frame.len) {
+      *status = HAL_ERR_CANSessionMux_InvalidBuffer;
+      return;
+    }
+    // TODO this time needs to be fixed up.
+    *timeStamp = msg.timestamp;
+    memcpy(data, msg.frame.data, msg.frame.len);
+    *dataSize = msg.frame.len;
+    *status = 0;
+    msg.timestamp = 0;
+    return;
+  }
+
+  // Add support for slow lookup later
+  *status = HAL_ERR_CANSessionMux_NotAllowed;
   return;
 }
 void HAL_CAN_OpenStreamSession(uint32_t* sessionHandle, uint32_t messageID,
