+++ hal/src/main/native/systemcore/CANInternal.h
