+++ hal/src/main/native/systemcore/HAL.cpp
@@ -28,11 +28,11 @@
 #include "HALInitializer.h"
 #include "HALInternal.h"
 #include "SystemServer.h"
-#include "SocketCan.h"
 #include "hal/DriverStation.h"
 #include "hal/Errors.h"
 #include "hal/Notifier.h"
 #include "hal/handles/HandlesInternal.h"
+#include "CANInternal.h"
 
 using namespace hal;
 
@@ -79,7 +79,6 @@ void InitializeHAL() {
   InitializeSerialPort();
   InitializeSPI();
   InitializeThreads();
-  InitializeSocketCan();
 }
 }  // namespace init
