+++ hal/src/main/native/systemcore/HALInitializer.h
@@ -52,5 +52,4 @@ extern void InitializeRelay();
 extern void InitializeSerialPort();
 extern void InitializeSPI();
 extern void InitializeThreads();
-extern void InitializeSocketCan();
 }  // namespace hal::init
